+++ QoL/HuntersMark.lua
@@ -30,57 +30,17 @@ HM.isPreview = false
 --------------------------------------------------------------------------------
 -- Inline helpers (KE has no core ApplyZoom/AddBorders/CreateIconFrame)
 --------------------------------------------------------------------------------
-local function ApplyZoom(texture, zoom)
-    local texMin = 0.25 * zoom
-    local texMax = 1 - 0.25 * zoom
-    texture:SetTexCoord(texMin, texMax, texMin, texMax)
-end
-
-local function AddBorders(frame, color)
-    local r, g, b, a = color[1], color[2], color[3], color[4] or 1
-
-    local top = frame:CreateTexture(nil, "OVERLAY", nil, 7)
-    top:SetHeight(1)
-    top:SetPoint("TOPLEFT", frame, "TOPLEFT", 0, 0)
-    top:SetPoint("TOPRIGHT", frame, "TOPRIGHT", 0, 0)
-    top:SetColorTexture(r, g, b, a)
-    top:SetTexelSnappingBias(0)
-    top:SetSnapToPixelGrid(false)
-
-    local bottom = frame:CreateTexture(nil, "OVERLAY", nil, 7)
-    bottom:SetHeight(1)
-    bottom:SetPoint("BOTTOMLEFT", frame, "BOTTOMLEFT", 0, 0)
-    bottom:SetPoint("BOTTOMRIGHT", frame, "BOTTOMRIGHT", 0, 0)
-    bottom:SetColorTexture(r, g, b, a)
-    bottom:SetTexelSnappingBias(0)
-    bottom:SetSnapToPixelGrid(false)
-
-    local left = frame:CreateTexture(nil, "OVERLAY", nil, 7)
-    left:SetWidth(1)
-    left:SetPoint("TOPLEFT", frame, "TOPLEFT", 0, 0)
-    left:SetPoint("BOTTOMLEFT", frame, "BOTTOMLEFT", 0, 0)
-    left:SetColorTexture(r, g, b, a)
-    left:SetTexelSnappingBias(0)
-    left:SetSnapToPixelGrid(false)
-
-    local right = frame:CreateTexture(nil, "OVERLAY", nil, 7)
-    right:SetWidth(1)
-    right:SetPoint("TOPRIGHT", frame, "TOPRIGHT", 0, 0)
-    right:SetPoint("BOTTOMRIGHT", frame, "BOTTOMRIGHT", 0, 0)
-    right:SetColorTexture(r, g, b, a)
-    right:SetTexelSnappingBias(0)
-    right:SetSnapToPixelGrid(false)
-end
+-- Icon helpers: KE:ApplyIconZoom() and KE:AddIconBorders() in Core/Widgets.lua
 
 local function CreateIconFrame(parent, iconSize)
     local frame = CreateFrame("Frame", nil, parent)
     frame:SetSize(iconSize, iconSize)
 
-    AddBorders(frame, { 0, 0, 0, 1 })
+    KE:AddIconBorders(frame)
 
     frame.icon = frame:CreateTexture(nil, "ARTWORK")
     frame.icon:SetAllPoints(frame)
-    ApplyZoom(frame.icon, 0.3)
+    KE:ApplyIconZoom(frame.icon)
 
     function frame:SetIconSize(newSize)
         self:SetSize(newSize, newSize)